feat(coverage): add incompleteBatchCount to domain coverage cells
Exposes an incompleteBatchCount field per coverage cell that counts
non-aggregate runs whose transcript count is below expected
(scenarioSelections × models × samplesPerScenario). The web matrix
renders an amber dot on affected cells and shows the count in the
popover with a note that not all transcripts were generated.

Also splits domain-coverage.ts into a companion gql-types file to
keep both under the 400-line limit.
